reverted tests and added new
