+++ tests/Form/Type/SecurityRolesMatrixTypeTest.php
@@ -13,24 +13,22 @@ declare(strict_types=1);
 
 namespace Sonata\UserBundle\Tests\Form\Type;
 
-use Sonata\UserBundle\Form\Type\SecurityRolesType;
+use Sonata\UserBundle\Form\Type\SecurityRolesMatrixType;
 use Sonata\UserBundle\Security\EditableRolesBuilder;
+use Sonata\UserBundle\Security\RolesMatrixBuilder;
 use Symfony\Component\Form\Extension\Core\Type\ChoiceType;
 use Symfony\Component\Form\FormTypeInterface;
 use Symfony\Component\Form\PreloadedExtension;
 use Symfony\Component\Form\Test\TypeTestCase;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
-/**
- * @author Quentin Fahrner <quentfahrner@gmail.com>
- */
-class SecurityRolesTypeTest extends TypeTestCase
+class SecurityRolesMatrixTypeTest extends TypeTestCase
 {
     protected $roleBuilder;
 
     public function testGetDefaultOptions(): void
     {
-        $type = new SecurityRolesType($this->roleBuilder);
+        $type = new SecurityRolesMatrixType($this->roleBuilder);
 
         $optionResolver = new OptionsResolver();
         $type->configureOptions($optionResolver);
@@ -45,7 +43,7 @@ class SecurityRolesTypeTest extends TypeTestCase
 
     public function testGetParent(): void
     {
-        $type = new SecurityRolesType($this->roleBuilder);
+        $type = new SecurityRolesMatrixType($this->roleBuilder);
 
         $this->assertEquals(ChoiceType::class, $type->getParent());
     }
@@ -79,46 +77,42 @@ class SecurityRolesTypeTest extends TypeTestCase
         $this->assertNull($form->getData());
     }
 
-    public function testSubmitWithHiddenRoleData(): void
+    public function testChoicesAsValues(): void
     {
-        $originalRoles = ['ROLE_SUPER_ADMIN', 'ROLE_USER'];
+        $resolver = new OptionsResolver();
+        $type = new SecurityRolesMatrixType($this->roleBuilder);
 
-        $form = $this->factory->create($this->getSecurityRolesTypeName(), $originalRoles, [
-            'multiple' => true,
-            'expanded' => true,
-            'required' => false,
-        ]);
+        // If 'choices_as_values' option is not defined (Symfony >= 3.0), default value should not be set.
+        $type->configureOptions($resolver);
 
-        // we keep hidden ROLE_SUPER_ADMIN and delete available ROLE_USER
-        $form->submit([0 => 'ROLE_USER']);
+        // If 'choices_as_values' option is defined (Symfony 2.8), default value should be set to true.
+        $resolver->setDefault('choices_as_values', true);
+        $type->configureOptions($resolver);
+        $options = $resolver->resolve();
 
-        $this->assertNull($form->getTransformationFailure());
-        $this->assertTrue($form->isSynchronized());
-        $this->assertCount(2, $form->getData());
-        $this->assertContains('ROLE_SUPER_ADMIN', $form->getData());
+        $this->assertTrue($resolver->hasDefault('choices_as_values'));
+        $this->assertTrue($options['choices_as_values']);
     }
 
     protected function getExtensions()
     {
-        $this->roleBuilder = $roleBuilder = $this->createMock(EditableRolesBuilder::class);
+        $this->roleBuilder = $roleBuilder = $this->createMock(RolesMatrixBuilder::class);
 
-        $this->roleBuilder->expects($this->any())->method('getRoles')->will($this->returnValue([
-            'ROLE_FOO' => 'ROLE_FOO',
-            'ROLE_USER' => 'ROLE_USER',
-            'ROLE_ADMIN' => 'ROLE_ADMIN: ROLE_USER',
+        $this->roleBuilder->expects($this->any())->method('getAllRoles')->will($this->returnValue([
+          'ROLE_FOO' => 'ROLE_FOO',
+          'ROLE_USER' => 'ROLE_USER',
+          'ROLE_ADMIN' => 'ROLE_ADMIN: ROLE_USER',
         ]));
 
-        $this->roleBuilder->expects($this->any())->method('getRolesReadOnly')->will($this->returnValue([]));
-
-        $childType = new SecurityRolesType($this->roleBuilder);
+        $childType = new SecurityRolesMatrixType($this->roleBuilder);
 
         return [new PreloadedExtension([
-            $childType->getName() => $childType,
+          $childType->getName() => $childType,
         ], [])];
     }
 
     private function getSecurityRolesTypeName()
     {
-        return 'Sonata\UserBundle\Form\Type\SecurityRolesType';
+        return 'Sonata\UserBundle\Form\Type\SecurityRolesMatrixType';
     }
 }
